Change calico-containers tagging schema
Since calico-containers is a downstream job it can have the same
git-sha for different upstreams, e.g:

  calico    --> calico-containers ---> img+gitsha1
  libcalico --> calico-containers ---> img+gitsha1

we have the same images naming since no changes in calico-containers
itself. To fix the problem we add GERRIT_CHANGE_NUMBER which is uniq.
At the same time, on merge event and manuall triggering we are setting:
    GERRIT_CHANGE_NUMBER=mcp-0.1

BUILD="${GERRIT_CHANGE_NUMBER:-mcp-0.1}-$(git rev-parse --short HEAD)"

PS. The same is true for calico-cni

Change-Id: Ibcf8077a838ec5319dd3c33b35ef0f6baccefc11