Do not build Calico CNI with latest 'libcalico-go'
Currently we are taking the latest commit from 'mcp' branch
of 'libcalico-go' go while building Calico CNI binaries.
That approach doesn't provide us a lot of benefits, but at
the same time it causes code sync (with upstream) issues,
because the library is often developed much earlier than
project which use it.

Change-Id: I7794379aacd447f2c5987808cd59a5fd8869717a