Pin kargo version for all system tests
Currently Kargo 'master' branch is broken [1]
and the fix [2] can't be merged due to CI issues.
In order to unblock our Calico and Kubernetes
tests jobs we need to pin Kargo version (last
working commit).

Generally, it's not safe to rely on 'master' of
Kargo, so we need re-pin to the git tag (>2.0.0)
when it's released).

[1] https://github.com/kubernetes-incubator/kargo/issues/647
[2] https://github.com/kubernetes-incubator/kargo/pull/648

Change-Id: I830f761cf66581f6accd8fb7fe5563c5c2b66aa4