Add Decapod project
Decapod is intendend to simplify deployment and lifecycle management of
Ceph.

Using this tool, it is possible to deploy clusters with best known
practices, add new nodes to cluster, remove them and purge cluster
if not needed anymore. It provides simple API to manage cluster
configurations. Also, it is possible to use it web UI to manage your
clusters in several clicks.

Decapod uses Ansible with ceph-ansible community project to deliver
the best user experience. Every action is supported by plugins, which
encapsulate best practices and settings. If user do not like proviided
settings, it is always possible to customize configuration before
execution.

https://github.com/Mirantis/ceph-lcm
http://decapod.readthedocs.io/en/latest/

Change-Id: I55a9c594a44edb91caa54e04cb752edd0ea5758c