Fix search of downloaded VM image
When you just added a slave to jenkins and runs job,
it downloads image and sets its name ("img" variable)
to packer/ubuntu-16.04*.qcow2. As result it makes
incorrect link to image file. For all other runs
on the same slave the "img" variable value is full
path to the image, so link to the image is correct.

Change-Id: I33ea219e456108dc6b710f19666c3d91837e8edb