Add new pipeline for CCP Mitaka to Newton upgrade
This will be a periodically called job to test if CCP can be upgraded from
Mitaka to Newton correctly. It uses (and depends on) CCP configuration with
reduced number of OpenStack services (as not all are supported in upgrades).
It is a simple configuration, meant to be stable and with minimal
dependencies, not using any frameworks, etc. just plain ccp + configs.

Change-Id: Ia8ba12bdb4bac5d0d566f8e5246c9995ecd4d0d0
Depends-On: I2f2b7a869a4d45705a35e08046a0b774930c36e7
Depends-On: I525ab291b1211cc0354d4cb30b5a7aa28a5d009a