Added project-config-base and project-config-local
To properly separate repositories and jobs deployed on-site for
clint-side CI it is needed to create project-configs for base (the
ones needed both for on-site and centralized CI/CDs) and for local
(used to manage actual deployments onsite) repositories and jobs.

To bootstrap this repositories appropriate entries are added to the
old project-config. After the change is implemented they will be
removed, since the new repositories will contain their own gerrit
records.

Change-Id: I41c224a5342884784bbebbeb9f93e0a283699e78