Added temporary jobs to watch project-config-base
Requires https://gerrit.mcp.mirantis.net/#/c/1268/ to be merged first.

The jobs to check commits to jenkins jobs and gerrit repositories
configurations are now defined in project-config-base repository.

Since the existing job-builder of mcp-ci does not watch
project-config-base repository it is needed to bootstrap the jobs
which will do that to created the new jobs.
To do that a "tmp" group of jobs are added. They are temporary jenkins
check/update jobs which are practically identical to the new jobs of
project-config-base repo. They will be used for bootstrapping only and
then will be removed from here.

Change-Id: I73d2a0de7eb4039c0bbe1733bf75e750b6303dd0